requirements.txt
pyspark==3.5.1
delta-spark==3.1.0
mlflow==2.16.2
apache-airflow==2.10.2
pandas==2.2.2
pyarrow==16.1.0
scikit-learn==1.5.2
torch==2.4.1
